Split to a new models package
